base-foc: introduce thread garbage collector
Instead of directly deleting a thread kernel-object when a thread
gets destroyed via core's CPU session, we collect their references
and let a dedicated thread do the actual deletion. This is to
prevent deletion of a thread and its timeslice when potentially
running on the same, which led to problems on Fiasco.OC in the
past.

Ref #1495
Ref #2659
Ref #2848