fixup "nic_router: multiple uplinks" (handle_link_state)
When the link state of an uplink interface changes, do not discard the
IP config of the corresponding domain in general. This is a remnant of
the times when the uplink was domain and interface in one, but now, each
domain might have multiple uplinks and downlinks. So, on link state
changes, the IP config shall be discarded only if all interfaces of the
domain are down.

Issue #2840