intel/gpu: pipe control write to PPHSWP index 0x34
Post sync write (QW_WRITE) is needed to wait for cache flushes to
finsh. When GLOBAL_GTT_IVB is not set and STORE_DATA_INDEX is set the
address in word 1 is interpreted as offset from beginnin Per process
hardware status page. We write to word 0x34 (reserved) as Linux does.