fixup "lx_emul: add basic ARM 32-bit support" (compiler-gcc.h)
The header compiler-gcc.h is not arch specific in Linux and so shouldn't be the
shadow header of us. This commit was triggered by the WireGuard port on
Arm_v8a. The port failed to compile after a rebase because the arm_64 version
of compiler-gcc.h tried to include the Linux header with
'#include_next <linux/compiler-gcc.h>'. However, this included
arm/lx_emul/shadow/linux/compiler-gcc.h instead because WireGuard
has both include paths src/include/spec/arm_64 and src/include/spec/arm
(which is correct).

Ref #4499