pc/audio: wrap 'usleep_range_state' instead of 'usleep_range'
'usleep_range' has become a static inline function in kernel 6.1 which
we cannot wrap. The function calls 'usleep_range_state' instead, which
we can wrap. We have to wrap this function because the SOF drivers calls
'usleep_range' in the microsecond range. The original code schedules a
timeout, which we don't support in sub-millisecond range, therefore,
we implement it using 'udelay' for sound.