lx_emul/shadow/page_alloc: add support for compound pages
Compound pages are a range of pages that are treated as one memory
range, which will be released completely if the compound head page is
released. Compound pages are allocated with the __GFP_COMP flag. This
concept is currently adapted into the "folio" concept, where a tail page
(which is any page following the head page) can be used to release the
whole memory range. A page can be cast into a folio.

This commit fills in the meta-data required for compound pages for the
'__alloc_pages' function.