libc: rework pthread_once
pthread_once used to allocate one pthread_mutex per
pthread_once_t-context. Because the mutex member of pthread_once_t is
opaque to the caller, the mutex was not freed upon release of
pthread_once_t.

With this commit we use one static recursive pthread_mutex in order to
avoid dynamic allocations which might not be freed later. The drawback
of this solutions is when a init_once call blocks, a potential other
thread using another once-context may block forever.

improves #5336