hw/x86: use tsc for keeping kernel time
For compatibility, we are using the APIC timer in one-shot mode
for scheduling. Since the timer stops counting until rescheduled,
it is not suitable for time keeping and lead to drifting kernel time
on different cores. By using the TSC instead of a one-shot timer for
calculating kernel time, we circumvent this effect.

Issue #5240
Issue #5144