libc: binary_name return pointer to local variable
The stripping of the path in the binary_name function returned a local
variable on the stack. Add 'with_binary_name' function to delay
destruction of the Path object till after Shared_object has been
created. Note that 'name' argument of the Shared_object will be copied
in the linker.

issue #5709