test-smp: remove false-positive in affinity test
This patch removes the spinning thread of the boot cpu within the
affinity test to circumvent competition with the main measurement
thread, and false-positive results of a dead thread on cpu 0.
Moreover, it shifts the de-blocking mechanism of all spinning
threads behind their initial LOG message, which again prevents
false-positives of a bunch of threads waiting at core's LOG
service running on the boot cpu, concurrently to the active
measurement main thread.

Fix #5696