libc: soften dlopen check constraint
Checking for a '.lib.so' suffix prevents unintentional loading of
non shared-objects, e.g. 'libdep.a' that prompted this change in the
first place and is in line with how libraries are conventionally
named on Genode.

It, however, makes it more cumbersome to deal with plugins produced
by other build-systems that commonly only feature a '.so' suffix
when not explicitly altered. So dropping the '.lib' portion remedies
this situation but still guards against the original issue.

Issue #5595.